New Functions and adding a field to the students api

+++ database/emergencyContact.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Date, Integer, ForeignKey, String, Table
+from sqlalchemy import Column, Date, Integer, ForeignKey, String, Table, BigInteger
 from dependencies import Base
 
 
@@ -7,7 +7,7 @@ class EmergencyContact(Base):
     id = Column(Integer, primary_key=True)
     name = Column(String(50), nullable=False)
     last_name = Column(String(50), nullable=False)
-    phone_mobile = Column(Integer, nullable=False)
-    phone_home = Column(Integer)
+    phone_mobile = Column(BigInteger, nullable=False)
+    phone_home = Column(BigInteger)
     relationship = Column(String(20), nullable=False)
     student = Column(Integer, ForeignKey('students.id'))
